Create README.md
The very first page of this project is the home page where the user can see the post titles and the date they were created on while they tap on that link it will take the user to the detailed page of the post.There was a few options on navbar also like about,contact,register,login,logout and sample post (Authenticated).Here one user can successfully register with their details and login to their profile to post some of the blogs they want to write. This post requests will go to admin panel where admin will read the posts and delete the post already posted on the site. There was another option to manage the about section on UI side admin can CRUD the about section from admin panel. Admin also can call people who tried to get connected with the admin.